requirements.txt
absl-py
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
async-timeout
attrs
certifi
charset-normalizer
click
contourpy
cycler
dataclasses-json
distro
exceptiongroup
faiss-cpu
fastapi
ffmpy
filelock
fonttools
frozenlist
fsspec
gradio
gradio-client
h11
httpcore
httpx
huggingface-hub
idna
importlib-resources
jinja2
jiter
joblib
jsonpatch
jsonpointer
kiwisolver
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-openai
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
mpmath
multidict
mypy-extensions
networkx
nltk
numpy
openai
orjson
packaging
pandas
pillow
propcache
pydantic
pydantic-core
pydub
pygments
pyparsing
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
regex
requests
requests-toolbelt
rich
rouge-score
ruff
safetensors
scikit-learn
scipy
semantic-version
sentence-transformers
shellingham
six
sniffio
sqlalchemy
starlette
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
tomlkit
torch
tqdm
transformers
typer
typing-extensions
typing-inspect
tzdata
urllib3
uvicorn
websockets
yarl
zipp
